Initial commit
Creates necessary files to get started with this Flask project. All
files in 'static' and 'templates' directories were obtained from
'Initializr'. Python files are empty ('__init__.py' is necessary for
this to be a valid package, and 'thermos.py' will contain the Flask
project). Filled README with all information so far. Ignoring '.idea'
and the virtual environment directory.